Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-    float a;
-    scanf("%lf",&a);
-    printf("You entered: %.4lf",a);
+    double a;
+    scanf("%f",&a);
+    printf("You entered: %.4f",a);
 
 }
